+++ src/segment.cpp
@@ -6,7 +6,7 @@ segment::segment() :
     sprite {{ tile_size,tile_size }}
 {
     sprite.setFillColor(sf::Color::Green);
-    
+
     sprite.setPosition(-200,-200);
 }
 
@@ -19,13 +19,13 @@ void segment::set_pos(int x, int y)
 {
     pos.x = x;
     pos.y = y;
-    sprite.setPosition({ pos.x * tile_size, pos.y * tile_size });
+    sprite.setPosition({ static_cast<float>(pos.x * tile_size), static_cast<float>(pos.y * tile_size) });
 }
 
 void segment::set_pos(sf::Vector2i new_pos)
 {
     pos = new_pos;
-    sprite.setPosition({ pos.x * tile_size, pos.y * tile_size });
+    sprite.setPosition({ static_cast<float>(pos.x * tile_size), static_cast<float>(pos.y * tile_size) });
 }
 
 sf::Vector2i segment::get_pos()
